requirements.txt
# Core ML / RAG
torch>=2.0.0
transformers>=4.38.0
peft>=0.9.0
sentence-transformers>=2.7.0
rank-bm25>=0.2.2

# Vector DB
pinecone>=3.0.0

# Code parsing
tree-sitter>=0.23.0
tree-sitter-python>=0.23.0

# LLM APIs
groq>=0.37.0
google-genai>=1.0.0

# LangChain / LangGraph
langchain>=1.2.0
langchain-core>=1.2.0
langchain-groq>=1.1.0
langgraph>=1.1.0

# Evaluation & data
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
streamlit>=1.30.0
